+++ nupic/embodied/models/__init__.py
@@ -19,3 +19,7 @@
 #  http://numenta.org/licenses/
 #
 # ------------------------------------------------------------------------------
+from .multiheaded_sparse_mlp import MultiHeadedSparseMLP
+from .multiheaded_dendrite_mlp import MultiHeadedDendriticMLP
+from .dendrite_mlp import DendriticMLP
+from .sparse_mlp import SparseMLP
